+++ api/middleware/__init__.py
@@ -1,0 +1,9 @@
+from api.middleware.correlation_id import LogCorrelationIdMiddleware
+from api.middleware.onion import OnionHostMiddleware
+from api.middleware.prometheus import PrometheusMiddleware
+
+__all__ = [
+    "LogCorrelationIdMiddleware",
+    "OnionHostMiddleware",
+    "PrometheusMiddleware",
+]
